backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
djangorestframework-jsonapi>=4.0.0
psycopg2-binary>=2.9.1
markdown
django-filter
python-dotenv
Pillow>=10.0.0
Faker>=10.0.0
django-crispy-forms
crispy-bootstrap5
django-allauth
django-cors-headers
djangorestframework-simplejwt
dj-rest-auth>=2.1,<3.0